+++ eir-fmm/src/reference_count/mod.rs
@@ -2,6 +2,8 @@ mod clone;
 mod drop;
 mod pointers;
 mod record_utilities;
+mod records;
 
 pub use clone::*;
 pub use drop::*;
+pub use records::*;
